Only deploy if there are changes

+++ deploy.sh
@@ -8,12 +8,11 @@ set -e
 
 rm -rf .docz/dist
 mkdir -p .docz/dist
-
 git clone https://${GITHUB_TOKEN}@github.com/KovuTheHusky/developer.myfursona.com.git --branch gh-pages .docz/dist
-
 yarn build
-
 cd .docz/dist
-git add --all
-git commit -a -m "Travis #$TRAVIS_BUILD_NUMBER"
-git push --force origin gh-pages
+if [ -n "$(git diff --quiet)" ]; then
+    git add --all
+    git commit -a -m "Travis #$TRAVIS_BUILD_NUMBER"
+    git push --force origin gh-pages
+fi
